EDITOR-#173 index.js to ts
