+++ controller-server/package.json
@@ -1,20 +1,23 @@
 {
   "name": "controller-server",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "index.ts",
   "license": "MIT",
-  "type": "module",
   "dependencies": {
     "@types/websocket": "^1.0.5",
     "body-parser": "^1.19.1",
     "express": "^4.17.2",
+    "typescript": "^4.5.5",
     "ws": "^8.4.2"
   },
   "scripts": {
-    "start": "nodemon ./index.js"
+    "start": "nodemon ./index.ts"
   },
   "devDependencies": {
+    "@types/express": "^4.17.13",
+    "@types/ws": "^8.5.1",
     "nodemon": "^2.0.15",
-    "prettier": "^2.5.1"
+    "prettier": "^2.5.1",
+    "ts-node": "^10.5.0"
   }
 }
