+++ controller-server/type/index.ts
@@ -1,0 +1,2 @@
+export interface TargetClientList { [key: string]: any } // any need to change to socket type later
+
